Sudoku Solver
This is a Sudoku Puzzle Solver written in C. It takes one line of input per puzzle in the form "8..912....3....4" with 81 characters in total (no spaces in between) with empty cells represented by a '.'; it reads characters until it reaches 81, if there are more it returns an invalid puzzle error message. It not only solves the puzzle but will also detect any invalid puzzles that are entered. It uses a brute force recursive back-tracking algorithm, it's functional and effective but not as efficient as I would like it to be. I'm playing around with a bit-manipulation method that represents a row or column as a single 32-bit number which is masked, shifted, etc. as necessary to represent valid cells for a number.  But I wanted to share this implementation first.