Bubble Shooter Game
This is a weekend project that I took on to refresh my Java skills. It's a basic Bubble Shooter game with a unique GUI, it's fairly straightforward but the physics need a lot of work as I didn't have time to work out the correct Trigonometry implementations for the ricochets. Will upload a fixed version when time allows.